Add defensive check for IsDiposed when getting texture (#1041)
Added additional condition to check if the `_whitePixelTexture` has been disposed of before reusing.  This prevents `ObjectDisposedExceptions` from occurring when using the texture.  An `ObjectDisposedException` can occur in the rare event of a `GraphicsDeviceReset` event trigger.  Additionally on Android, if the application is placed in the background for too long, the android activity may stop but the application will still be running.  If this occurs, when switching back to the application, the texture may be disposed of.