Add World Bounds Constraint to Orthographic Camera (#1049)
* Mark Pitch as obsolete

Pitch was introduced in PR #907. While the intentions of the implementation were good, moving forward it is going to start causing technical debt to keep it as a property for all camera types. Having pitch doesn't make sense for an Orthographic camera, and would be better served as an implementation into an isometric or even a perspective camera implementation in the future.

This is marked as a compiler warning (not an error), so existing code will continue to work. The property will be removed in the next major version.

* Add WorldBounds property to OrthographicCamera

Implements #64
Adds the ability to constrain camera movement to a specified rectangular area.

* Mark PitchUp and PitchDown as obsolete

Pitch was introduced in PR #907. While the intentions of the implementation were good, moving forward it is going to start causing technical debt to keep it as a property for all camera types. Having pitch doesn't make sense for an Orthographic camera, and would be better served as an implementation into an isometric or even a perspective camera implementation in the future.

This is marked as a compiler warning (not an error), so existing code will continue to work. The property will be removed in the next major version.

* Add missing XML documentation

* Reorganize file according to coding guidelines

* Add missing XML documentation

* Remove old Camera2DTests file, no longer used

* Add extension for assert matrix equality with precision

* Add tests for OrthographicCamera

* Zoom should only clamp to world bounds if explicitly told to do so

* Use the default 800x480 MonoGame resolution for testing

* Make `IsClampedToWorldBounds` an auto property

* Call site should check if methods should execute

* Renamed IsClampedToWorldBounds

* Rename method to use Clamp instead of Limit

* Simplify by using ToVector2

* Remove old code

* Only clamp if required

* No need to throw exception since we clamp anyway

* Use ArgumentOutOfRangeException helper methods

* Clamp pitch on setting

* Properties handle clamping now, no need for dedicated methods

* Remove extra new line

* Update tests for changes

* Update workflow to use xvfb for testing